LIBITD-879. Implemented eligibility authorization
Added "lending_org_code" and "auth_org_code" to session cookies.

Callback page now looks for an "eduPersonScopedAffiliation" property
from Shibboleth of the form "member@<auth_org_code>.edu". If the
property is found, the user is considered authorized. If not, the
user is considered not authorized.

Updated view to show whether the user was authorized. Removed
"Authentication Successful" banner, because the user will only
ever get to the page if the authentication is successful, and it
may have been confused with an authorization.

https://issues.umd.edu/browse/LIBITD-879