LIBITD-2409. Updated Rubocop to latest version
Updated Rubocop to the latest version. Added "rubocop-rails" as
"Rails" cops have been split off into their own gem.

Updated "rubocop.yml" for latest version, and added some additonal
exceptions to better fit the needs of this project (such as disabling
the "Rails/ActionControllerTestCase" cop), and to be consistent with
other SSDR Rails projects (setting "Style/EmptyMethod" to "expanded").

Disabled the "Style/FrozenStringLiteralComment" cop because it turns
out the frozen string literals by default were never added to Ruby 3,
it is unclear whether it will ever be added in the future, and this
project never had the comment to begin with.

https://umd-dit.atlassian.net/browse/LIBITD-2409