LIBITD-2421. Modified test to eliminate .env interference
Modified the test to use "assert_includes" and the banner text, instead
of the full HTML "div", because the presences of a ".env" file with
banner environment variables could cause the test to fail (as they
would get picked up and included in the div).

Not a problem for the other tests, which more narrowly define the
environment variables.

https://umd-dit.atlassian.net/browse/LIBITD-2421