LIBITD-2589. Pin "base64" gem version for Kubernetes Passenger Phusion
Pinned "base64" gem version to 0.1.0 due to the following error
when running the application in Kubernetes using Passenger Phusion:

```
You have already activated base64 0.1.1, but your Gemfile requires
base64 0.3.0.
```

This is apparently because "base64" is a "default" gem provided by
Ruby, and Passenger is loading it first.

In the "Gemfile.lock", the "base64" gem is loaded by the "rack-session"
gem, which has "base64 (>= 0.1.0)" for the dependency, so using the
0.1.1 version should be okay.

This customization should be revisited when upgrading the Ruby version.

https://umd-dit.atlassian.net/browse/LIBITD-2589