LIBITD-2589. Pin "logger" gem version for dev Passenger Phusion
Pinned "logger" gem version to 1.5.3 due to the following error
when running the application in the
"umd-lib/reciprocal-borrowing-dev-env" development environment

```
You have already activated logger 1.5.3, but your Gemfile requires
logger 1.7.0.
```

This is apparently because "logger" is a "default" gem provided by
Ruby, and Passenger is loading it first.

In the "Gemfile.lock", the "logger" gem is loaded by the "activesupport"
gem, which has "logger (>= 1.4.2)" for the dependency, so using the
1.5.3 version should be okay.

This customization should be revisited when upgrading the Ruby version.

https://umd-dit.atlassian.net/browse/LIBITD-2589